requirements/Cloud9.txt
asn1crypto==0.22.0
attrs==17.2.0
Automat==0.6.0
backports-abc==0.5
beautifulsoup4==4.6.0
bleach==2.0.0
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
constantly==15.1.0
cryptography==2.0.3
cssselect==1.0.1
decorator==4.1.2
Django==1.11.4
django-cors-headers==2.1.0
django-filter==1.0.4
djangorestframework==3.6.4
entrypoints==0.2.3
html5lib==0.999999999
hyperlink==17.3.0
idna==2.5
incremental==17.5.0
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.1.0
jupyter-core==4.3.0
livereload==2.5.1
lxml==3.8.0
Markdown==2.6.9
MarkupSafe==1.0
mistune==0.7.4
mkdocs==0.16.3
mkdocs-material==1.10.1
mysqlclient==1.3.10
nbconvert==5.2.1
nbformat==4.3.0
notebook==5.0.0
numpy==1.13.1
pandas==0.20.3
pandocfilters==1.4.2
parsel==1.2.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyasn1==0.3.2
pyasn1-modules==0.0.11
pycparser==2.18
PyDispatcher==2.0.5
Pygments==2.2.0
pymdown-extensions==4.0
pyOpenSSL==17.2.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.1
queuelib==1.4.2
requests==2.18.3
service-identity==17.0.0
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.13
terminado==0.6
testpath==0.3.1
tornado==4.5.1
traitlets==4.3.2
Twisted==17.5.0
typing==3.6.2
urllib3==1.22
w3lib==1.18.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==2.0.0
xlrd==1.1.0
zope.interface==4.4.2
